+++ src/Exception/ConcurrencyLimitExceededException.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * Hyvä Themes - https://hyva.io
+ * Copyright © Hyvä Themes. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+declare(strict_types=1);
+
+namespace Hyva\Ai\Exception;
+
+use Magento\Framework\Exception\LocalizedException;
+
+class ConcurrencyLimitExceededException extends LocalizedException
+{
+}